+++ Mile.HyperV/Mile.HyperV.Guest.Protocols.h
@@ -16,6 +16,7 @@
 //   - MsvmPkg\VmbusDxe\ChannelMessages.h
 //   - MsvmPkg\Include\Vmbus\VmbusPacketFormat.h
 //   - MsvmPkg\VideoDxe\SynthVidProtocol.h
+//   - MsvmPkg\Include\Protocol\SynthKeyProtocol.h
 
 #ifndef MILE_HYPERV_GUEST_PROTOCOLS
 #define MILE_HYPERV_GUEST_PROTOCOLS
@@ -807,6 +808,55 @@ const GUID HK_CONTROL_CLASS_ID =
     0x48EA,
     { 0xBD, 0x65, 0xF9, 0x27, 0xA6, 0x1C, 0x76, 0x84 }
 };
+
+#define HK_MAKE_VERSION(Major, Minor) \
+    ((HV_UINT32)(Major) << 16 | (HV_UINT32)(Minor))
+#define HK_VERSION_WIN8 HK_MAKE_VERSION(1, 0)
+
+typedef enum _HK_MESSAGE_TYPE
+{
+    HkMessageProtocolRequest = 1,
+    HkMessageProtocolResponse = 2,
+    HkMessageEvent = 3,
+    HkMessageSetLedIndicators = 4,
+} HK_MESSAGE_TYPE;
+
+typedef struct _HK_MESSAGE_HEADER
+{
+    HK_MESSAGE_TYPE MessageType;
+} HK_MESSAGE_HEADER, *PHK_MESSAGE_HEADER;
+
+typedef struct _HK_MESSAGE_PROTOCOL_REQUEST
+{
+    HK_MESSAGE_HEADER Header;
+    HV_UINT32 Version;
+} HK_MESSAGE_PROTOCOL_REQUEST, *PHK_MESSAGE_PROTOCOL_REQUEST;
+
+typedef struct _HK_MESSAGE_LED_INDICATORS_STATE
+{
+    HK_MESSAGE_HEADER Header;
+    HV_UINT16 LedFlags;
+} HK_MESSAGE_LED_INDICATORS_STATE, *PHK_MESSAGE_LED_INDICATORS_STATE;
+
+typedef struct _HK_MESSAGE_PROTOCOL_RESPONSE
+{
+    HK_MESSAGE_HEADER Header;
+    HV_UINT32 Accepted : 1;
+    HV_UINT32 Reserved : 31;
+} HK_MESSAGE_PROTOCOL_RESPONSE, *PHK_MESSAGE_PROTOCOL_RESPONSE;
+
+typedef struct _HK_MESSAGE_KEYSTROKE
+{
+    HK_MESSAGE_HEADER Header;
+    HV_UINT16 MakeCode;
+    HV_UINT32 IsUnicode : 1;
+    HV_UINT32 IsBreak : 1;
+    HV_UINT32 IsE0 : 1;
+    HV_UINT32 IsE1 : 1;
+    HV_UINT32 Reserved : 28;
+} HK_MESSAGE_KEYSTROKE, *PHK_MESSAGE_KEYSTROKE;
+
+#define HK_MAXIMUM_MESSAGE_SIZE 256
 
 // *****************************************************************************
 // Microsoft Hyper-V SCSI Controller